subprojects/wdm/requirements.txt
pydantic
pyyaml
fire
